added  Userauthentication to add & delete expense

+++ controllers/expense.js
@@ -1,5 +1,6 @@
 
 const { where } = require('sequelize');
+
 const Expense=require('../models/expense');
 
 function isStringInvalid(string){
@@ -10,7 +11,7 @@ function isStringInvalid(string){
     }
 }
 
-exports.addexpense=async(req,res,next)=>{
+exports.addExpense=async(req,res,next)=>{
     try{
         const {amount,date,reason,category}=req.body;
         if(isStringInvalid(amount)||isStringInvalid(date)||
@@ -18,8 +19,10 @@ exports.addexpense=async(req,res,next)=>{
         {
             return res.status(500).json({message:'Bad Parmeters:Something is Missing',success:false})
         }
-       const response= await Expense.create({amount,date,reason,category})
-        res.status(200).json({message:response,success:true});
+       /* const response= await Expense.create({amount,date,reason,category,userId:req.user.id});  */
+       const response= await req.user.createExpense({amount,date,reason,category});
+        
+        res.status(201).json({message:response,success:true});
         
     }catch(err){
         res.status(500).json({message:"Something went wrong",success:false})
@@ -30,11 +33,11 @@ exports.addexpense=async(req,res,next)=>{
 exports.getExpenses=async(req,res,next)=>{
 
     try{
-        const response=await Expense.findAll();
+        /* const response =await Expense.findAll({where:{userId:req.user.id}}); */
+        const response =await req.user.getExpenses();
         res.status(200).json({message:response,success:true});
     }
     catch(err){
-
         res.status(500).json({message:err,success:true});
 
     }
@@ -44,7 +47,11 @@ exports.deleteExpense=async(req,res,next)=>{
 
     try{
         const id=req.params.id;
-        const response=await Expense.destroy({where:{id:id}});
+       if(isStringInvalid(id))
+       {
+        return  res.status(500).json({message:'something went wrong',success:false})
+       }
+        const response=await Expense.destroy({where:{id:id,userId:req.user.id}});
         res.status(200).json({message:response,success:true});
     }
     catch(err){
